requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
idna==3.10
linkify-it-py==2.0.3
markdown-it-py==3.0.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
platformdirs==4.3.6
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
textual==0.85.2
typer==0.13.0
typing_extensions==4.12.2
uc-micro-py==1.0.3
urllib3==2.2.3
